requirements.txt
fastapi
uvicorn
openai
httpx
python-docx
pydantic-settings
python-multipart
faiss-cpu
numpy
sentence-transformers
python-dotenv
loguru
